add atan2()  returns values between -pi and pi. when working in radians, you can add 2π or subtract 2π until the angle is within the desired range.
atan2() returns values between -pi and pi. When calculating phi in y = z - h(x) for radar measurements, the resulting angle phi in the y vector should be adjusted so that it is between -pi and pi. The Kalman filter is expecting small angle values between the range -pi and pi. HINT: when working in radians, you can add 2π or subtract 2π until the angle is within the desired range.